requirements.txt
PyAutoGUI~=0.9.54
tesserocr~=2.6.1
jellyfish~=1.0.1
requests~=2.31.0
Pillow~=10.0.1
matplotlib~=3.8.0
mss~=9.0.1
PyYAML~=6.0.1
pynput~=1.7.6